Renamed utils folder to helpers to avoid conflict when merging to main
