+++ helpers/timer.js
